Avoid HTTP 302 Redirect on Root Rest Endpoint
In OpenHAB 3.1 the endpoint is ``http(s)://openhab:80/rest/`` instead of ``http(s)://openhab:80/rest``.

In my environment, this behavior has caused an error. Openhab redirected from HTTPS to HTTP. This caused a "400 Bad Request" on nginx.

This tweak avoids that redirect. I don't know if this also works with openhab 2.x.